+++ apps/server/src/server.ts
@@ -5,8 +5,7 @@ import { serve, setup } from 'swagger-ui-express';
 
 import { errorHandler, provider } from '@overture-stack/lyric';
 
-import { appConfig } from './config/server.js';
-import { getServerConfig } from './config/server.js';
+import { appConfig, getServerConfig } from './config/app.js';
 import swaggerDoc from './config/swagger.js';
 import healthRouter from './routes/health.js';
 import pingRouter from './routes/ping.js';
